backend/requirements.txt
Pillow
qwen-vl-utils==0.0.8
fastapi
uvicorn
requests
python-multipart
python-dotenv
beautifulsoup4
